refractoring frontend
