+++ frontend/src/services/leetcode_review/constants.js
@@ -1,0 +1,21 @@
+export const PREDEFINED_PROBLEM_CATEGORIES = [
+  { name: 'Arrays & Hashing', count: 0 },
+  { name: 'Linked List', count: 0 },
+  { name: 'Tree', count: 0 },
+  { name: 'Graph', count: 0 },
+  { name: '1D DP', count: 0 },
+  { name: 'Binary Search', count: 0 },
+  { name: 'Stack', count: 0 },
+  { name: '2D DP', count: 0 },
+  { name: 'Bit Manipulation', count: 0 },
+  { name: 'Two Pointers', count: 0 },
+  { name: 'Sliding Window', count: 0 },
+  { name: 'Math', count: 0 },
+  { name: 'Matrix', count: 0 },
+  { name: 'Intervals', count: 0 },
+  { name: 'Heap', count: 0 },
+  { name: 'Advanced Graph', count: 0 },
+  { name: 'Backtracking', count: 0 },
+];
+
+export const LEETCODE_API_URL = process.env.REACT_APP_LEETCODE_API_URL;
