+++ frontend/src/services/leetcode_review/utils.js
@@ -1,0 +1,14 @@
+export const formatDate = (timestamp) => {
+  if (!timestamp) return 'Not Reviewed';
+  const date = new Date(timestamp);
+  return date.toLocaleDateString('en-US', {
+    weekday: 'short',
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit',
+    second: '2-digit',
+    timeZoneName: 'short',
+  });
+};
