+++ frontend/src/views/leetcode/Review/ProblemCategoriesTable.js
@@ -8,7 +8,6 @@ import {
   TableHead,
   TableRow,
   Paper,
-  Checkbox,
   FormControl,
   InputLabel,
   Select,
@@ -18,8 +17,11 @@ import {
   Typography,
 } from '@mui/material';
 import { AuthContext } from 'contexts/authContext';
-
-const leetcodeAPIURL = process.env.REACT_APP_LEETCODE_API_URL;
+import CustomCheckbox from '../common/CustomCheckbox';
+import { fetchAllLeetcodeQuestionsNonBlocking } from 'services/leetcode_review/apiFetchAllLeetcodeQuestions';
+import { updateProblemReviewCategories } from 'services/leetcode_review/apiUpdateProblemReviewCategories';
+import ProgressBar from '../common/ProgressBar';
+import SubmitButton from '../common/SubmitButton';
 
 const ProblemCategoriesTable = ({
   userSubmissionsByReviewCategory,
@@ -31,46 +33,15 @@ const ProblemCategoriesTable = ({
 }) => {
   const { uuid } = useContext(AuthContext);
   const [filter, setFilter] = useState('All');
-  const [allProblemData, setAllProblemData] = useState([]); // To hold all problem data
+  const [, setAllProblemData] = useState([]); // To hold all problem data
 
   useEffect(() => {
     // Fetch problem data once when the component mounts
-    const fetchProblemData = async () => {
-      try {
-        const response = await fetch(`${leetcodeAPIURL}/problems/all`);
-        const data = await response.json();
-        setAllProblemData(data);
-      } catch (error) {
-        console.error('Error fetching problem data', error);
-      }
-    };
-
-    fetchProblemData();
+    fetchAllLeetcodeQuestionsNonBlocking(setAllProblemData);
   }, []);
 
   const handleSubmit = async () => {
-    try {
-      let problemCategorySet = [];
-      for (const category of problemCategoriesMarkedForReview) {
-        problemCategorySet.push(category);
-      }
-
-      const payload = { category: problemCategorySet };
-      await fetch(
-        `${leetcodeAPIURL}/users/${uuid}/problem/categories/review/submit`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(payload),
-        }
-      );
-      alert('Submitted successfully');
-    } catch (error) {
-      console.error('Error submitting data', error);
-      alert('Failed to submit');
-    }
+    updateProblemReviewCategories(uuid, problemCategoriesMarkedForReview);
   };
 
   const filteredProblemCategories =
@@ -79,30 +50,6 @@ const ProblemCategoriesTable = ({
       : userSubmissionsByReviewCategory.filter((category) =>
           problemCategoriesMarkedForReview.includes(category.name)
         );
-
-  console.log('filteredProblemCategories:', filteredProblemCategories);
-  console.log('userProblemSubmissions:', userProblemSubmissions);
-  const calculateProgress = (category) => {
-    console.log('category:', category);
-    const problems = userProblemSubmissions.filter(
-      (problem) => problem.category === category.name
-    );
-
-    console.log('problems:', problems);
-
-    const completed = problems.filter(
-      (problem) =>
-        !problem.next_review_timestamp ||
-        new Date(problem.next_review_timestamp) >=
-          new Date(problem.last_reviewed_timestamp)
-    ).length;
-
-    return {
-      completed,
-      total: category.count,
-      progress: category.count ? (completed / category.count) * 100 : 0,
-    };
-  };
 
   return (
     <div>
@@ -149,8 +96,6 @@ const ProblemCategoriesTable = ({
           </TableHead>
           <TableBody>
             {filteredProblemCategories.map((category) => {
-              const { completed, total, progress } =
-                calculateProgress(category);
               return (
                 <TableRow
                   key={category.name}
@@ -163,56 +108,27 @@ const ProblemCategoriesTable = ({
                   }}
                 >
                   {editMode && (
-                    <TableCell padding='checkbox'>
-                      <Checkbox
-                        checked={problemCategoriesMarkedForReview.includes(
-                          category.name
-                        )}
-                        onChange={() => onCheckboxChange(category.name)}
-                      />
-                    </TableCell>
+                    <CustomCheckbox
+                      checked={problemCategoriesMarkedForReview.includes(
+                        category.name
+                      )}
+                      onChange={() => onCheckboxChange(category.name)}
+                    />
                   )}
                   <TableCell onClick={() => onTypeClick(category.name)}>
                     {category.name}
                   </TableCell>
-                  <TableCell>
-                    <Box display='flex' alignItems='center'>
-                      <Box minWidth={35}>
-                        <Typography
-                          variant='body2'
-                          color='textSecondary'
-                        >{`${completed} / ${total}`}</Typography>
-                      </Box>
-                      <Box width='100%' ml={1}>
-                        <LinearProgress
-                          variant='determinate'
-                          value={progress}
-                          sx={{
-                            '& .MuiLinearProgress-barColorPrimary': {
-                              backgroundColor:
-                                progress === 100 ? 'green' : 'orange',
-                            },
-                          }}
-                        />
-                      </Box>
-                    </Box>
-                  </TableCell>
+                  <ProgressBar
+                    userProblemSubmissions={userProblemSubmissions}
+                    category={category}
+                  />
                 </TableRow>
               );
             })}
           </TableBody>
         </Table>
       </TableContainer>
-      {editMode && (
-        <Button
-          variant='contained'
-          color='primary'
-          onClick={handleSubmit}
-          sx={{ marginTop: '16px' }}
-        >
-          Submit
-        </Button>
-      )}
+      {editMode && <SubmitButton handleSubmit={handleSubmit} />}
     </div>
   );
 };
